modifico el tp

+++ calculadora.h
@@ -5,6 +5,6 @@
 
 int add(int x, int y);
 int substract(int x, int y);
-int divide(int x, int y);
+float divide(int x, int y);
 int multiply(int x, int y);
 int factorial(int x);
